feat: update scoring system to use Double for points across various entities and services

+++ backend/src/main/java/com/tutor_management/backend/modules/exercise/dto/response/ExerciseListItemResponse.java
@@ -43,7 +43,7 @@ public class ExerciseListItemResponse {
     /**
      * Earned score if the submission has been graded.
      */
-    private Integer studentTotalScore;
+    private Double studentTotalScore;
 
     /**
      * Manual projection constructor used for optimized JPQL Repository queries.
