+++ backend/src/main/java/com/tutor_management/backend/modules/submission/dto/response/SubmissionListItemResponse.java
@@ -23,9 +23,9 @@ public class SubmissionListItemResponse {
     private String studentId;
     private String studentName;
     private SubmissionStatus status;
-    private Integer mcqScore;
-    private Integer essayScore;
-    private Integer totalScore;
+    private Double mcqScore;
+    private Double essayScore;
+    private Double totalScore;
     private LocalDateTime submittedAt;
     private LocalDateTime gradedAt;
 }
